requirements.txt
streamlit==1.29.0
openai>=1.0.0
python-dotenv==1.0.0
tweepy==4.14.0
supabase==2.3.1